refactor(root): moved error pages to root level components
