+++ src/components/ErrorPage/parts/HeadImage/index.tsx
