+++ src/components/ErrorPage/parts/HeadImage/styled.tsx
