+++ src/features/ui/index.ts
@@ -2,7 +2,6 @@ export * from './components/AccessibleHidden'
 export * from './components/AccountInfo'
 export * from './components/Button'
 export * from './components/Container'
-export * from './components/ErrorPage'
 export * from './components/Header'
 export * from './components/Input'
 export * from './components/LayoutExternal'
